Menus and UI
I've created a main menu that transitions to the TestArena when NewGame is clicked.

I also created UI that appears over a unit's token when it is selected, or disappears when deselected.  This is a big step; now I just need to figure out how to process inputs how I want and it will be a playable game.