Delegates and Talents
Finished up the framework for using the Delegates with Talents.

CombatController and MovementController will be going away, we're switching back to a class-based action handler.  This is easier since it also lets us dual-purpose the action instances for the Event Log.

Right now I'm getting an error with the delegates, though.  it says something about object being null.  I'll have to see what's up with that.