Talent Framework complete
Converted all the delegate methods to void since it's just easier to do it that way.  Allows for easy null checks and streamlines the code drastically.  Since we pass the AttackData as as parameter we can just have the listener methods do their own thing on the data.

May change the TalentTrigger name to AbilityTrigger to make it more generic.  This would allow us to use the same framework for weapon abilities.