Trying to re-organize menu control
It's meh.  Initially I wanted to move away from having a MenuController and just do separate scripts for each one but it turns out you can't do that if you're using a script on that object when it's inactive...

So now I'm right back where I started except I'm instancing a prefab of the menu.  It won't delete the menu when you deselect but I think that might actually be because I'm an idiot and didn't call the right Event in deselect.  Duh.

Now I just need to figure out how to tie the button clicks back into game functionality.  Since they're being instanced, they'll need to dynamically get new listeners or what not.