Equip Fully Functional
Finished handling the Equip functionality.  I haven't tested with the edge cases involving the talents using the hook-ins, but I'm hoping it shouldn't be too hard to implement.

I'm keeping MenuController, according to the MVC pattern.  The Controller has a hard reference to an Action, and when a new Action is created it will set that hard reference to itself.

I've decided on trying out a "Camp" style management scene instead of just boring menus.  This will allow me to re-use code from the existing controllers.