Trigger Generation, Talent Learning
Implemented Trigger Generation.  By using a foreach loop on the unit's talent list, we can use a try block to look for a Trigger class by the same name.  If we find it with Type.GetType, then we addcomponent.  Otherwise, a TypeNotFound exception is handled in the catch block.

Also, I adjusted the menu for learning/viewing talents.  Uses vertical lists with buttons now.  not the best though, since button clicks don't allow you to pass dynamic parameters.  Why unity whyyyy.

Also looked into using a scroll view, but that shit is weirdly hard.  It'll scroll stuff just fine, but it won't mask things so you're just moving the panel up and down.