Talent Menu Work
Updated the UI, fixed a lot of things.  Added the last functionality.  It now properly displays and loads talents.  The menu even checks for talents that are already learned.

All that's left is to create a method linked to the Buy button that checks against whether it's learned, then sends a buy flag.